Update MultiMarkdown to 6.4.0

+++ ext/mmd/uuid.c
@@ -102,8 +102,8 @@ char * uuid_string_from_bits(unsigned char * raw) {
 	char * result = malloc(37);
 
 	sprintf(result, "%02x%02x%02x%02x-%02x%02x-%02x%02x-%02x%02x-%02x%02x%02x%02x%02x%02x",
-	        raw[0], raw[1], raw[2], raw[3], raw[4], raw[5], raw[6], raw[7],
-	        raw[8], raw[9], raw[10], raw[11], raw[12], raw[13], raw[14], raw[15] );
+			raw[0], raw[1], raw[2], raw[3], raw[4], raw[5], raw[6], raw[7],
+			raw[8], raw[9], raw[10], raw[11], raw[12], raw[13], raw[14], raw[15] );
 
 	return result;
 }
@@ -149,6 +149,6 @@ void custom_seed_rand(void) {
 	// This is not a "cryptographically secure" random seed,
 	// but good enough for an EPUB id....
 	unsigned long seed = mix(clock(), time(NULL), clock());
-	srand(seed);
+	srand((unsigned int)seed);
 }
